move ImageFeatures MatchInfo and CameraParams in core/misc/
