+++ modules/python/src2/cv2.cpp
@@ -210,9 +210,6 @@ typedef std::vector<KeyPoint> vector_KeyPoint;
 typedef std::vector<Mat> vector_Mat;
 typedef std::vector<std::vector<Mat> > vector_vector_Mat;
 typedef std::vector<UMat> vector_UMat;
-typedef std::vector<detail::ImageFeatures> vector_ImageFeatures;
-typedef std::vector<detail::MatchesInfo> vector_MatchesInfo;
-typedef std::vector<detail::CameraParams> vector_CameraParams;
 typedef std::vector<DMatch> vector_DMatch;
 typedef std::vector<String> vector_String;
 typedef std::vector<Scalar> vector_Scalar;
